+++ needle/modules/hooking/frida/script_dump-keychain.py
@@ -5,9 +5,9 @@ class Module(FridaScript):
     meta = {
         'name': 'Frida Script: Keychain Dumper',
         'author': 'Bernard Wagner (@MWRLabs)',
-        'description': 'Retrieve all keychain items',
+        'description': 'Retrieve all the keychain items belonging to the target application',
         'options': (
-
+            ('output', True, False, 'Full path of the output file'),
         ),
     }
 
@@ -73,7 +73,15 @@ if (ObjC.available) {
 } else {
     console.log("Objective-C Runtime is not available!");
 }
-        '''
+'''
+
+    # ==================================================================================================================
+    # UTILS
+    # ==================================================================================================================
+    def __init__(self, params):
+        FridaScript.__init__(self, params)
+        # Setting default output file
+        self.options['output'] = self.local_op.build_output_path_for_file("frida_script_dump_keychain.txt", self)
 
     # ==================================================================================================================
     # RUN
@@ -81,7 +89,6 @@ if (ObjC.available) {
     def module_run(self):
         # Run the payload
         try:
-            self.results = []
             self.printer.info("Parsing payload")
             hook = self.JS
             script = self.session.create_script(hook)
@@ -89,16 +96,8 @@ if (ObjC.available) {
             script.load()
         except Exception as e:
             self.printer.warning("Script terminated abruptly")
-            print(e)
-
-    def on_message(self, message, data):
-        try:
-            if message:
-                self.results.append(json.loads(message["payload"]))
-        except Exception as e:
-            print(message)
-            print(e)
+            self.printer.warning(e)
 
     def module_post(self):
-        for key in self.results:
-            print(json.dumps(key, indent=4, sort_keys=True))
+        self.printer.info("Keychain Items:")
+        self.print_cmd_output()